refactor: typed out SideNavLink
