--- packages/react/src/components/UIShell/SideNavLink.js
@@ -1,110 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2023
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import cx from 'classnames';
-import PropTypes from 'prop-types';
-import React, { useContext } from 'react';
-import Link, { LinkPropTypes } from './Link';
-import SideNavIcon from './SideNavIcon';
-import SideNavItem from './SideNavItem';
-import SideNavLinkText from './SideNavLinkText';
-import { usePrefix } from '../../internal/usePrefix';
-import { SideNavContext } from './SideNav';
-
-const SideNavLink = React.forwardRef(function SideNavLink(
-  {
-    children,
-    className: customClassName,
-    renderIcon: IconElement,
-    isActive,
-    isSideNavExpanded,
-    large = false,
-    tabIndex,
-    ...rest
-  },
-  ref
-) {
-  const isRail = useContext(SideNavContext);
-
-  const prefix = usePrefix();
-  const className = cx({
-    [`${prefix}--side-nav__link`]: true,
-    [`${prefix}--side-nav__link--current`]: isActive,
-    [customClassName]: !!customClassName,
-  });
-
-  return (
-    <SideNavItem large={large}>
-      <Link
-        {...rest}
-        className={className}
-        ref={ref}
-        tabIndex={
-          tabIndex === undefined
-            ? !isSideNavExpanded && !isRail
-              ? -1
-              : 0
-            : tabIndex
-        }>
-        {IconElement && (
-          <SideNavIcon small>
-            <IconElement />
-          </SideNavIcon>
-        )}
-        <SideNavLinkText>{children}</SideNavLinkText>
-      </Link>
-    </SideNavItem>
-  );
-});
-
-SideNavLink.displayName = 'SideNavLink';
-SideNavLink.propTypes = {
-  ...LinkPropTypes,
-
-  /**
-   * Specify the text content for the link
-   */
-  children: PropTypes.node.isRequired,
-
-  /**
-   * Provide an optional class to be applied to the containing node
-   */
-  className: PropTypes.string,
-
-  /**
-   * Specify whether the link is the current page
-   */
-  isActive: PropTypes.bool,
-
-  /**
-   * Property to indicate if the side nav container is open (or not). Use to
-   * keep local state and styling in step with the SideNav expansion state.
-   */
-  isSideNavExpanded: PropTypes.bool,
-
-  /**
-   * Specify if this is a large variation of the SideNavLink
-   */
-  large: PropTypes.bool,
-
-  /**
-   * Provide an icon to render in the side navigation link. Should be a React class.
-   */
-  renderIcon: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-
-  /**
-   * Optional prop to specify the tabIndex of the button. If undefined, it will be applied default validation
-   */
-  tabIndex: PropTypes.number,
-};
-
-// eslint-disable-next-line react/display-name
-export const createCustomSideNavLink = (element) => (props) => {
-  return <SideNavLink element={element} {...props} />;
-};
-
-export default SideNavLink;
